Remove UpdateRender() call from OverlayForm_Load
UpdateRender() will destroy the Browser window, which was created due to Navigate
causing the addon to see a valid Browser but then destroy it out from under the
addon on the Cef thread without it knowing that it is even coming.

Instead only UpdateRender() from Navigate() which the addon can overload and
observe happening.

This also avoids loading the url twice on startup, once from setting the Url
and once from the form load afterward.