Store the current CefBrowser as a single field and make access thread-safe
There is only 1 active CefBrowser at any time, created when UpdateRender() calls to
Renderer::BeginRender(). The BeginRender() method will destroy any existing browsers
then make a new one.

The CefBrowser is not known from the addon main thread in BeginRender() as it is created
asynchronously on the Cef UI thread, which then calls to Renderer::OnCreated() once it
exists. We use a semaphore to safely store this CefBrowser in a private field
|Renderer.browser|, and use the same semaphore to access it from the public read-only
field |Renderer.Browser|.

When destroying the browser, we set the |Renderer.browser| private field, using the
semaphore, back to null. This gives us the following invariant:
- If you check that Renderer.Browser is not null on the main thread it will stay
non-null while you are in a task on that thread.

This allows code to do if (Browser != null) { .. use Browser .. } safely. Note that
you can't rely on the variable to stay null, as it is changed to non-null from the
Cef UI thread, but this is not problematic as when it's null there is nothing to use.

Since we're setting the browser back to null on the main thread, we don't need to do
any action in OnBeforeClose().