Remove double-navigate on startup
The InitializeOverlay() method in OverlayBase will Navigate to the
Config.Url to be loaded. Then it also sets a timer for 1s later to
again Navigate to Config.Url. This has the effect of loading
everything twice, which is more work, and causing everything to
flash about 1 second after starting.