Add a way for javascript to call back to the overlay plugin.
This adds an overlayMessage() function to the OverlayPluginApi object,
which will call a virtual OverlayMessage() function on OverlayBase.
Plugins can override it and parse the string to do whatever they wish,
such as Text-to-Speech for example.

The sendMessage() function takes 2 arguments, which are the |name| of
the overlay plugin to send it to, and the |message| string to send.
overlayMessage() takes the same 2 arguments.

In order to send the overlayMessage() to the same plugin which is
running the javascript, the name of the plugin needs to be known,
but it is user defined. So we also add an |overlayName| readonly
string to the OverlayPluginApi object, that has the name of the
overlay plugin. Then the javascript may do the following to send
a string to its c# plugin:

OverlayPluginApi.overlayMessage(OverlayPluginApi.overlayName, 'hello');