Add overlay resources for target bars
There are four html pages that are all the same, other than a
differently named container.  The JavaScript finds the container by
name to figure out which target bar type it's supposed to be.
Settings for each bar are stored separately, so they can all be
configured differently.

Much of the settings code is cribbed from cactbot.

The settings UI is pretty lackluster looking, but maybe somebody
else could improve it.  Sorry!

Followups here could be including some of the enmity values
(time to death, enmity, relative enmity) as possible options.