Add options for ending encounters
This adds two opt-in event settings for ending encounters, one on wipe
and one five seconds (non-configurable) after game combat ends.

This is already implemented in both cactbot and Triggernometry, and so
it makes some sense to consolidate these options in one place where
they can be useful to everybody.

The "five seconds" combat option is intended for the use cases of
older fights (such as ARR/HW where the Network6D line is not sent)
and dungeons (to split before a boss).