Improve EffectiveDistance precision in target bars
EffectiveDistance is just a (rounded) byte and so is not that useful
for determining exact distances from boss mechanics.  This PR
changes EffectiveDistance to be a float based on the Radius when
on 6.x.  For backwards compatibility, pre-6.x users will get the
rounded byte.

Belated thanks to @KattTails for the memory location for radius!