Update URL for Kagami (#4)
Recreating this PR for this fork per the message on discord
Original PR: ngld#266

Original Message:
The original URL was for a fork that has since been deleted. The fork is removed and the old URL now returns a 404 error.
The proposed URL is for another fork which has been maintained by @sgosiaco, and has been generally recommended in the ACT_FFXIV Discord.