Update OverlayPlugin to support newer ffxiv plugin versions (#12)
Remove Language.Unknown and calls to ToProperCase.

Additionally, stop proper casing 6.x versions as the ffxiv plugin has
stopped proper casing names there.  This caused problems in cactbot
where names no longer matched, so doing it here as well for consistency.

Split from ngld/OverlayPlugin#235 to make this easier to land.
This is a copy of ngld/OverlayPlugin#258 applied to this fork.